// src/endpoints.d.ts
export declare enum Endpoints {
    info = "info",
    volume = "volume",
    speaker = "speaker",
    key = "key",
    nowPlaying = "now_playing",
    select = "select",
    sources = "sources",
    getZone = "getZone",
    setZone = "setZone",
    addZoneSlave = "addZoneSlave",
    removeZoneSlave = "removeZoneSlave",
    bassCapabilities = "bassCapabilities",
    bass = "bass",
    presets = "presets",
    name = "name",
    getGroup = "getGroup"
}